Update main.js
Tried changing the algorithm to search all paths based on the condition that you can take any path as long as the next street's weight is between 1 and K more than the current streets weight (1 <= Y - X <= K). And the next street can be any street that we haven't used. Still no dice.